predict triggers

+++ lib/models/nn.py
@@ -1141,7 +1141,7 @@ class NN(Configurable):
     trigger_indices = tf.cast(tf.where(tf.equal(srl_targets, trigger_label_idx)), tf.int32)
     idx = tf.stack([trigger_indices[:,0], trigger_indices[:,2]], -1)
 
-    targets = tf.scatter_nd(idx, tf.ones([tf.shape(idx)[0]]), [batch_size, bucket_size])
+    targets = tf.scatter_nd(idx, tf.ones([tf.shape(idx)[0]], dtype=tf.int32), [batch_size, bucket_size])
 
     cross_entropy = tf.nn.sparse_softmax_cross_entropy_with_logits(logits=logits, labels=targets)
     cross_entropy *= self.tokens_to_keep3D
